First Source Release by Arisa
Added missing function for reputation label
ReSharper cleanup code
Fixed one bug with wrong type of variable needed for one function (something with long, int etc.. there was one mistake, I make quick fix)

+++ src/Disqus.NET/Requests/DisqusPostCreateRequest.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Disqus.NET.Requests
 {
-    public class DisqusPostCreateRequest: DisqusRequestBase
+    public class DisqusPostCreateRequest : DisqusRequestBase
     {
         private DisqusPostCreateRequest(string message)
         {
@@ -15,7 +15,7 @@ namespace Disqus.NET.Requests
             return new DisqusPostCreateRequest(message);
         }
 
-        public DisqusPostCreateRequest Parent(int parentPostId)
+        public DisqusPostCreateRequest Parent(long parentPostId)
         {
             Parameters.Add(new KeyValuePair<string, string>("parent", parentPostId.ToString()));
